add confirmation to delete

+++ src/app/users/user.service.ts
@@ -1,9 +1,9 @@
+// userService.ts
 import { Injectable } from '@angular/core';
-import {Observable} from "rxjs";
-import {HttpClient} from "@angular/common/http";
-import {User} from "./user";
-import { map } from 'rxjs/operators';
-import {environment} from "../../environments/environment";
+import { Observable } from "rxjs";
+import { HttpClient } from "@angular/common/http";
+import { User } from "./user";
+import { environment } from "../../environments/environment";
 
 @Injectable({
   providedIn: 'root'
@@ -30,8 +30,19 @@ export class UserService {
     return this.http.put<User>(`${this.apiUrl}/users/${user.utilisateurId}`, user);
   }
 
-
   deleteUser(id: number): Observable<string> {
-    return this.http.delete(`${this.apiUrl}/users/${id}`, { responseType: 'text' });
-  }
+   
+    const isConfirmed = window.confirm('Are you sure you want to delete this user?');
+
+ 
+    if (isConfirmed) {
+        return this.http.delete(`${this.apiUrl}/users/${id}`, { responseType: 'text' });
+    } else {
+
+        return new Observable<string>((observer) => {
+            observer.next('Deletion canceled');
+            observer.complete();
+        });
+    }
+}
 }
